requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=3.1.1 # Apache-2.0
six>=1.11.0 # MIT
PyYAML>=3.13 # MIT
ansible>=2.8,!=2.8.9,!=2.9.12,<2.10.0
ansible-runner>=1.4.0 # Apache-2.0
cliff>=3.2.0 # Apache-2.0
